+++ client/components/Layout.tsx
@@ -1,4 +1,4 @@
-import React, { useCallback, useEffect, useReducer, useState } from "react";
+import React, { useCallback, useEffect, useState } from "react";
 import { Link } from "react-router-dom";
 import ReactForm from "../components/ReactForm";
 import useMeasure from "../hooks/useMeasure";