+++ server/controllers/posts.ts
@@ -1,7 +1,7 @@
 import SQL, { SQLStatement } from "sql-template-strings";
 import objectHash from "node-object-hash";
 import { PoolClient } from "pg";
-import { Post, PostRelation, PostSearchResults, PostSummary } from "../routes/apiTypes";
+import { Post, PostSearchResults, PostSummary } from "../routes/apiTypes";
 import * as db from "../helpers/db";
 import HTTPError from "../helpers/HTTPError";
 import { MIME_EXT, rangeRatingRegex } from "../helpers/consts";
@@ -133,9 +133,9 @@ export async function random(tag: string | null = null): Promise<PostSummary | n
     sample = SQL`(
       SELECT filtered.ids[floor(random() * icount(filtered.ids)) + 1] as id
       FROM (
-        SELECT union_agg(tags_postids.postids) AS ids
+        SELECT union_agg(tag_postids.postids) AS ids
         FROM tags
-        INNER JOIN tags_postids ON tags_postids.tagid = tags.id
+        INNER JOIN tag_postids ON tag_postids.tagid = tags.id
         WHERE tags.name LIKE ${pattern} OR tags.subtag LIKE ${pattern}
       ) filtered
     )`;
@@ -309,18 +309,18 @@ async function getCachedPosts(key: CacheKey, client?: PoolClient): Promise<Cache
     WITH
       whitelist AS (
         SELECT intersection_agg(ids) as ids FROM (
-          SELECT union_agg(tags_postids.postids) AS ids
+          SELECT union_agg(tag_postids.postids) AS ids
           FROM unnest(${whitelist}::TEXT[]) WITH ORDINALITY x(pat, patid)
           LEFT JOIN tags ON tags.name LIKE pat OR tags.subtag LIKE pat
-          INNER JOIN tags_postids ON tags_postids.tagid = tags.id
+          INNER JOIN tag_postids ON tag_postids.tagid = tags.id
           GROUP BY patid
         ) ids
       ),
       blacklist AS (
-        SELECT union_agg(tags_postids.postids) AS ids
+        SELECT union_agg(tag_postids.postids) AS ids
         FROM unnest(${blacklist}::TEXT[]) pat
         LEFT JOIN tags ON tags.name LIKE pat OR tags.subtag LIKE pat
-        INNER JOIN tags_postids ON tags_postids.tagid = tags.id
+        INNER JOIN tag_postids ON tag_postids.tagid = tags.id
       ),
       filtered AS `.append(filtered).append(SQL`
     SELECT