+++ server/helpers/dbImport/indexes.sql
@@ -5,18 +5,23 @@ ALTER TABLE urls ADD CONSTRAINT urls_postid_fkey FOREIGN KEY (postid) REFERENCES
 
 DELETE FROM mappings WHERE NOT EXISTS (SELECT 1 FROM posts WHERE id = postid);
 DELETE FROM mappings WHERE NOT EXISTS (SELECT 1 FROM tags WHERE id = tagid);
-DELETE FROM tags WHERE NOT EXISTS (SELECT 1 FROM mappings WHERE id = tagid);
+DELETE FROM tags WHERE NOT EXISTS (SELECT 1 FROM mappings WHERE mappings.tagid = tags.id)
+                   AND NOT EXISTS (SELECT 1
+                                   FROM mappings
+                                   INNER JOIN tag_siblings ON tag_siblings.betterid = mappings.tagid
+                                   WHERE tag_siblings.tagid = tags.id);
 ALTER TABLE mappings ADD CONSTRAINT mappings_postid_fkey FOREIGN KEY (postid) REFERENCES posts(id) ON DELETE CASCADE,
                      ADD CONSTRAINT mappings_tagid_fkey FOREIGN KEY (tagid) REFERENCES tags(id) ON DELETE CASCADE;
 CREATE INDEX ON mappings(tagid);
 CREATE INDEX tags_name_idx ON tags USING gin(name gin_trgm_ops);
 CREATE INDEX tags_subtag_idx ON tags USING gin(subtag gin_trgm_ops);
 
-CREATE MATERIALIZED VIEW tags_postids AS
+DROP TABLE IF EXISTS tag_postids;
+CREATE TABLE tag_postids AS
   SELECT tagid, sort(array_agg(postid)) AS postids
   FROM mappings
   GROUP BY tagid;
-CREATE UNIQUE INDEX ON tags_postids(tagid);
+CREATE UNIQUE INDEX ON tag_postids(tagid);
 
 CREATE INDEX ON posts(posted, id);
 CREATE INDEX ON posts(rating, id);