+++ server/helpers/dbImport/mappings.ts
@@ -1,23 +1,19 @@
-import { Writable } from "stream";
-import { Database, Statement } from "better-sqlite3";
+import { Database } from "better-sqlite3";
 import { PoolClient } from "pg";
-import { ServiceID } from "../consts";
-import { Import } from "./index";
+import { Import } from "./import";
 
-export default class Mappings extends Import<[number, number]> {
+export default class Mappings extends Import {
   display = "Mappings";
   
+  initialKey = [-1, -1];
   totalQuery = '';
   outputQuery = 'COPY mappings(postid, tagid) FROM STDIN (FORMAT CSV)';
   inputQuery = '';
   
-  constructor(hydrus: Database, postgres: PoolClient) {
+  constructor(hydrus: Database, postgres: PoolClient, service: number) {
     super(hydrus, postgres);
     
-    const service: { id: number } = hydrus.prepare(`SELECT service_id AS id FROM services WHERE service_type = ${ServiceID.LOCAL_TAG} LIMIT 1`).get();
-    if(!service) throw new Error("Unable to locate local tags service!");
-    
-    const mappingsTable = `current_mappings_${service.id}`;
+    const mappingsTable = `current_mappings_${service}`;
     
     this.inputQuery = `
       SELECT
@@ -32,18 +28,5 @@ export default class Mappings extends Import<[number, number]> {
     `;
     
     this.totalQuery = `SELECT count(1) FROM ${mappingsTable}`;
-  }
-  
-  async importBatch(lastKey: [number, number] | null, limit: number, input: Statement, output: Writable): Promise<[number, number] | null> {
-    const rows = input.all(lastKey === null ? -1 : lastKey[0], lastKey === null ? -1 : lastKey[1], limit);
-    
-    let buf = "";
-    for(const row of rows) {
-      buf += row[2];
-    }
-    output.write(buf);
-    
-    if(rows.length > 0) return [rows[rows.length - 1][0], rows[rows.length - 1][1]];
-    else return null;
   }
 }